Add a few more test cases.
Also includes changes to make these tests pass.  In particular, one test
has dead code (an exception handler that cannot possibly be reached).
The current translation eliminates this dead code (and the changes to
the analysis are to accommodate that).

One consequence is that some blocks are left that are redundant and
could be coalesced into a smaller number of larger blocks.  The dead
code introduced the division, and since it is gone the division is no
longer necessary.  Alternatively, we could just leave in the dead code...