Fix a significant class of SSA labeling bug.
In a few places, the labeling code was using the 'Instruction'
type as a Map key.  This is incorrect because Instructions do not
have identity and two 'Instruction's that have different SSA values
can actually compare as equal.  This commit changes everything to
refer to 'Instruction's only by the index into the 'Instruction'
vector.