Decode the raw data for the *-payload instructions
Before, these instructions stored raw data as `[Word16]`.  This change
decodes them properly*.  The switch payloads are decoded into `Int32`
offsets.  The array-data-payload is decoded such that all different
types of constant are promoted to Int64 (this includes float types,
since we can't know what the real type of a constant is until it is
used).  The decoding of actual Java longs might not be right - that
still requires a test or three.