Improve association of move-result with the relevant predecessor
Before, we were zipping instructions with their successors.
Unfortunately, only successors within the same basic block.  If we had
an invoke with a move-result in a different block (due to possible
exceptional control flow), it was missed.  This later caused a crash.
Instead, use a simpler solution based on instruction absolute
indices (now that we track those).  This is simpler and more correct.