Fix a subtle bug in trivial phi elimination
We are maintaining the mapping that records which basic blocks each
incoming phi value came from separately from the list of phi operands.
This makes trivial phi elimination easier.  Before this change, this
incoming block mapping was not being updated when a trivial phi was
removed.  This left a dangling reference that manifest as a failed map
lookup (and failure) later.

Now we are correctly replacing these phis.