Fix a bug in handling wide arguments to invokes
This was a crashing bug.  Wide values (longs and doubles) occupy two
registers.  Instructions refer to wide values by the first register in
the pair.  Except Invoke instructions.  Both registers for a wide
argument are explicitly passed to invoke instructions.

The code was trying to translate the second register using
getFinalValue, but we never create values for the second register (by
design).  The fix is to just drop the second register from each wide
pair from the argument list before translating it.