Fix a bug in insnCodeUnits
This broke when I fixed the decoding of array data.  The code unit count
was being derived from the old [Word16] payload in each payload
instruction.  When I changed the types of these payloads, the counts
were off.  This caused errors in the SSA translation process since the
code checking the payload instruction could get the wrong offset into
the instruction stream and find an invalid payload.

The switch payloads were also affected and fixed in this change.