Fix another bug in handling wide values
This one only seems to occur in larger methods where virtual registers
are reused.  Instructions that produce a result and operate on wide
registers actually clobber the named destination register and the one
after it.  We were only accounting for the named destination register,
leaving dangling labels in the register map.  These dangling labels were
incorrectly ending up in phi nodes and causing crashes when we tried to
translate them.

This change modifies the labeling code to clobber wide registers when
necessary.  It also re-uses the wide argument filtering code from the
other step of the SSA translation.  This way we don't assign invalid
labels to the second half of a wide register pair.