Fix try block range calculation
Try blocks are marked by a start offset into the vector of shorts and a
number of short units covered by the block.  The old method of
determining the actual instructions covered had problems with boundary
conditions.  Doing it properly requires a bit more care, and this patch
does so with a scan from the start, taking into account all of the
instruction 'short' units correctly.

This uncovered a bug in the test suite where I was expecting a possible
exception, but apparently the compiler guarantees there cannot be one
there.  Disabled that test since it isn't right.