Fix a crash due to incorrect debug info
We use debug information to provide names for method parameters if it is
available.  In cases where the debug information is incorrect, this can
cause a crash.  Here, debug information is present for methods, but the
list of debug information for method parameters is empty.  When zipped
with the list of Parameter values, this caused a truncation.  Later on,
this caused the SSA transformation to start with an incorrect register
assignment and panic.

The fix is to only use debug information for parameter names if the
number of debug entries matches the number of parameters.