Make sure we are getting *all* types in the DexFile type table
The old type table included all of the types in the type table of the
raw apk dex file.  These are only types referred to by name in the
code.  This is not actually all of the types we use, however.  We also
"generate" new types for things like array-get (which returns a value of
the inner type of the array).  In the case when this type was not
mentioned explicitly in the dalvik, it was not in our type table.

This commit adds a pass to collect these types that we synthesize.