Add a new helper: resolveStaticField
This function figures out which class a given static field really
resides in.  The name mentioned in the bytecode might be a subclass of
the real class, so we need to find the real definition.

This change required a slight tweak to the imports of the class
hierarchy analysis so that we can import the CHA from the Utils module
without introducing cycles.