Support Dalvik annotations
Currently, annotations on classes and methods are fully supported in
both the low level and SSA IRs.  Annotations on fields and parameters
are supported in the low level IR.

Annotations cover both user-provided annotations with the @class.method
syntax in Java, as well as compiler-generated annotations that record
full type signatures (with generic parameters).  There are also compiler
annotations that describe containment relationships for anonymous inner
classes.