Fix a serious SSA construction bug
This relates back to handling of redefinitions of classes.  Our policy
is to always prefer earlier definitions over later definitions.  We were
correctly handling classes and method references, but not method
definitions.

This would manifest as a crash when inspecting the method def of certain
InvokeDirect instruction after deserializing a .dalvik file.  The ID of
the method def would refer to a method definition that had been
overwritten by a newer definition.  When the RTS tried to tie the knot
after deserialization, it would look for the ID in the MethodID ->
Method table and find nothing (because the method had been discarded at
serialization time).  This would turn into an error whenever that field
was actually inspected.  In practice, we rarely use that field except
when serializing a .dalvik file.

It is a bit shocking that we never encountered this crash before.