Switch from zip-archive to zip-conduit
zip-archive doesn't support data descriptors for entries.  Some APKs
have this extra metadata, which causes zip-archive to crash.

zip-conduit supports the extra data and seems to be a bit more robust.
I'd like to use the 'zip' package instead, but it depends (almost
certainly incorrectly) on a newer version of base than we can support
due to our reliance on ghc 7.8.